+++ src/result.rs
@@ -8,7 +8,7 @@ use std::convert::{From, Into};
 
 use httparse;
 use mio;
-#[cfg(all(not(windows), feature="ssl"))]
+#[cfg(feature="ssl")]
 use openssl::ssl::error::SslError;
 
 use communication::Command;
@@ -47,7 +47,7 @@ pub enum Kind {
     /// Indicates a failure to schedule a timeout on the EventLoop.
     Timer(mio::TimerError),
     /// Indicates a failure to perform SSL encryption.
-    #[cfg(all(not(windows), feature="ssl"))]
+    #[cfg(feature="ssl")]
     Ssl(SslError),
     /// A custom error kind for use by applications. This error kind involves extra overhead
     /// because it will allocate the memory on the heap. The WebSocket ignores such errors by
@@ -111,7 +111,7 @@ impl StdError for Error {
             Kind::Encoding(ref err) => err.description(),
             Kind::Io(ref err)       => err.description(),
             Kind::Parse(_)          => "Unable to parse HTTP",
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Kind::Ssl(ref err)      => err.description(),
             Kind::Queue(_)          => "Unable to send signal on event loop",
             Kind::Timer(_)          => "Unable to schedule timeout on event loop",
@@ -123,7 +123,7 @@ impl StdError for Error {
         match self.kind {
             Kind::Encoding(ref err) => Some(err),
             Kind::Io(ref err)       => Some(err),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Kind::Ssl(ref err)      => Some(err),
             Kind::Custom(ref err)   => Some(err.as_ref()),
             _ => None,
@@ -182,7 +182,7 @@ impl From<Utf8Error> for Error {
     }
 }
 
-#[cfg(all(not(windows), feature="ssl"))]
+#[cfg(feature="ssl")]
 impl From<SslError> for Error {
     fn from(err: SslError) -> Error {
         Error::new(Kind::Ssl(err), "")