+++ src/stream.rs
@@ -3,9 +3,9 @@ use std::net::SocketAddr;
 
 use mio::{TryRead, TryWrite};
 use mio::tcp::TcpStream;
-#[cfg(all(not(windows), feature="ssl"))]
+#[cfg(feature="ssl")]
 use openssl::ssl::SslStream;
-#[cfg(all(not(windows), feature="ssl"))]
+#[cfg(feature="ssl")]
 use openssl::ssl::error::Error as SslError;
 
 use result::{Result, Error, Kind};
@@ -13,7 +13,7 @@ use result::{Result, Error, Kind};
 use self::Stream::*;
 pub enum Stream {
     Tcp(TcpStream),
-    #[cfg(all(not(windows), feature="ssl"))]
+    #[cfg(feature="ssl")]
     Tls {
         sock: SslStream<TcpStream>,
         negotiating: bool,
@@ -27,16 +27,15 @@ impl Stream {
         Tcp(stream)
     }
 
-    #[cfg(all(not(windows), feature="ssl"))]
+    #[cfg(feature="ssl")]
     pub fn tls(stream: SslStream<TcpStream>) -> Stream {
         Tls { sock: stream, negotiating: false }
     }
 
-    #[cfg(all(not(windows), feature="ssl"))]
+    #[cfg(feature="ssl")]
     pub fn is_tls(&self) -> bool {
         match *self {
             Tcp(_) => false,
-            #[cfg(all(not(windows), feature="ssl"))]
             Tls {..} => true,
         }
     }
@@ -44,7 +43,7 @@ impl Stream {
     pub fn evented(&self) -> &TcpStream {
         match *self {
             Tcp(ref sock) => sock,
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { ref sock, ..} => sock.get_ref(),
         }
     }
@@ -52,7 +51,7 @@ impl Stream {
     pub fn is_negotiating(&self) -> bool {
         match *self {
             Tcp(_) => false,
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { sock: _, ref negotiating } => *negotiating,
         }
 
@@ -61,7 +60,7 @@ impl Stream {
     pub fn clear_negotiating(&mut self) -> Result<()> {
         match *self {
             Tcp(_) => Err(Error::new(Kind::Internal, "Attempted to clear negotiating flag on non ssl connection.")),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { sock: _, ref mut negotiating } => Ok(*negotiating = false),
         }
     }
@@ -69,7 +68,7 @@ impl Stream {
     pub fn peer_addr(&self) -> io::Result<SocketAddr> {
         match *self {
             Tcp(ref sock) => sock.peer_addr(),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { ref sock, ..} => sock.get_ref().peer_addr(),
         }
     }
@@ -77,7 +76,7 @@ impl Stream {
     pub fn local_addr(&self) -> io::Result<SocketAddr> {
         match *self {
             Tcp(ref sock) => sock.local_addr(),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { ref sock, ..} => sock.get_ref().local_addr(),
         }
     }
@@ -88,7 +87,7 @@ impl TryRead for Stream {
     fn try_read(&mut self, buf: &mut [u8]) -> io::Result<Option<usize>> {
         match *self {
             Tcp(ref mut sock) => sock.try_read(buf),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { ref mut sock, ref mut negotiating } => {
                 match sock.ssl_read(buf) {
                     Ok(cnt) => Ok(Some(cnt)),
@@ -110,7 +109,7 @@ impl TryWrite for Stream {
     fn try_write(&mut self, buf: &[u8]) -> io::Result<Option<usize>> {
         match *self {
             Tcp(ref mut sock) => sock.try_write(buf),
-            #[cfg(all(not(windows), feature="ssl"))]
+            #[cfg(feature="ssl")]
             Tls { ref mut sock, ref mut negotiating } => {
 
                 *negotiating = false;